Issue #201: Extract the height statevector from vtr using a pointer.
A new function in vtr is created to point to the local height
statevector. Instead of redoing the read trial and calculate the
height in s2c, this function is used to access the height
statevector in s2c, before the horizontal interpolation. With this
method, there is no need to allocate arrays and do extra transposes
to re-read the trial field and calculate the height.

The memory usage and timing requirements remain similar to the master.