Issue #243: Avoid make copy of 2D heigh to reduce the timing for slant-path calculation.
In the slantprofilelatlon_mod, there is no need to copy the full 2D
height at each level for the lat/lon calculation. At each level,
pass the 3D height with the levelIndex to the heightBilinearInterp
s/r. This solution reduces the timing substantially.

Add the check to ensure the interpolated lat/lon is inside the hco
domain at the top of heightBilinearInterp s/r.