Issue #243: Use a separate s/r to handle rejection/modification of the lat/lon of header.
There is no need to check if lat/lon column of the header is inside
domain in 's2c_setupBilinearInterp' or 's2c_setupInterpInfo' s/r. In
s2c, rejection of the obs and manipulation of the header's lat/lon
column is handled inside the new s/r, checkLatLonObs. This s/r is
called after the slant-path calculation (before MPI communication)
in s2c_setupInterpInfo so the modification of the header lat/lon is
done only in one place in the code. If the obs is outside the
domain and rejectOutsideObs is requested (in nl), then the obs is
not assimilated and the lat/lon column of the header is set to the
mid-point of the domain. If the obs is outside the domain and
rejectOutsideObs is .false. (in tl/ad) and the lat/lon of a level
exceed the domain boundary, the lat/lon is moved to the edge of
domain.

In s2c_setupInterpInfo, the computation of the rotated lat/lon for
the winds is performed in a separate loop.

Some name changes are done for the arrays containing the lat/lon.

Add/subtract the positioning offset (1e-4) when moving the lat/lon
to the edge of the domain by adjusting the xpos/ypos. This is to
avoid going outside the domain when re-calculating the lat/lon from
the modified xpos/ypos due to round-off error.