Issue #243: latlonChecks resets the lat/lon to the edge of domain.
After slant-path calculation and before the lat/lon MPI
communication, latlonChecks rejects the header if
rejectOutsideObs=.true. (in nl) and the last level lat/lon is outside
the domain. If rejectOutsideObs=.false. (in tl/ad), then the lat/lon
at each level is checked and is reseted to the edge of the domain if
outside.

Add the positionOffsetXY where the check for lat/lon in domain is done.