Issue #243: Adding print statement to track lat/lon change of a problematic header.
Print statement are added to locate the header with unrealistic
lat/lon in tovs_nl_mod.

The lat/lon column of this headerIndex is printed out after doing
slant-path calculation before and after latlonChecks (to verify if
the latlonCheck does the job for resetting the outside to the domain
edge), and compared against the lat/lon of this header after the MPI
communication.

The listing shows that the lack of precision for re-calculating the
position from lat/lon causes the point to go beyond the edge of the
domain. One solution would be to increase the positionOffsetXY
value to prevent this instances.