Issue #243: Abort if the xpos/ypos could go beyond domain boundary due to precision issue.
In latlonChecks, if the lat/lon is moved to the edge of domain,
xpos/ypos are re-calculated from lat/lon. If the re-computed
xpos/ypos exceed the domain boundary, it aborts. Since xpos/ypos are
recalculated from the lat/lon later in the code, this is the last
check to ensure the xpos/ypos will not go beyond the domain boundary
due to single precision calculation.

The print statements to alert lat/lon exceeding the domain in
s2c_setupBilinearInterp is not needed anymore.

If the lat/lon is outside the domain, the calculation of the
interpolated height for slant-path calculation will also abort.