Issue #243: Adding the namelist nams2c to trigger slant-path calculation.
Adding the namelist nams2c to control the positionOffset and
doSlantPath variables in s2c_setupInterpInfo.

Modification done to s2c_setupInterpInfo for the no slant-path case
to work.

The MPI communication for footPrintRadius is moved outside the
doSlantPath if condition.

Since we do not re-compute the xpos/ypos during slant-path lat/lon
calculation, the positionOffset is not needed for the bilinear
height interpolation.