Issue #243: latlonChecks is done for all types of obs, regardless of doSlantpath.
Regardless of doing the slant-path or not, the latlonChecks for all
the observations should be done at all time.

Necessary changes to set rejectOutsideObs=.false. in the global grid
for nl. In global grid, latlonChecks will ONLY move the obs inside
the domain.

Remove the part of the code to force the lon be equal to its
original value, when moving the obs at the pole. For the nl,
latlonChecksAnlGrid will move the obs inside the analysis domain and
latlonChecks will move it more, if still outside.

Change the nml to avoid slant-path for some UnitTests.