Issue #243: Move the obs at the boundary after lat/lon MPI comm.
In global domain after latlonChecksAnlGrid, the remaining obs are all
assimilable. Instead of moving the obs in the latlonChecks s/r
before MPI comm, if the obs is outside the domain the xpos/ypos are
adjusted before the calculation of the interpolation weights (after
the lat/lon MPI comm). The positionOffsetXY is no longer needed,
since xpos/ypos are used directly. The lat/lon are not re-
calculated from the xpos/ypos and the shifting of lat/lon location
due to roundoff error is avoided.

The nams2c namelist determines the slant-path calculation in
nl/tlad separately. So the slant-path can be done only for the
background state.