Issue #243: Fix the lat/lon at the surface.
The slant-path lat/lon at the surface is fixed to the original
lat/lon so the interpolated height/pressure at surface does not
change, compared to the vertical column. This will reduce possible
RT error due to interpolation (RTTOV uses the pressure/height at
the surface as input to perform the vertical interpolation).

slant-path calculation is modified so 0 <= lonSlant <= 2*PI.