Issue #249: New LETKF test on xc40
Now there are two tests for LETKF, a small one (on ppp) and a
slightly bigger one (on xc40).

The OMA is now computed and put in obsSpaceData by calling
the obs operators with the ensemble mean analysis as input.
This required a new optional argument to inn_computeInnovation
to allow the specification of which obsSpaceData column will
receive the resulting O-H(x) values (default is obs_omp).

The number of timesteps for the analyses is no longer hard-
coded to 1. Instead it can equal 1 or the number of timesteps
chosen for the background state.

A few minor adjustments to filenames to make this fit easier
into the EnKF maestro suite. Also, adding Anl mean P0 and surface
GZ to all increment files (needed by GEM for IAU).