Issue #247: Generalization of the observation transform - part 1
The actual observation transforms (from wind speed and direction to u and v), previously hardwired, is now dynamically actived (or not) based on the list of readed and assimilated observations (no need to modify the namelist!).

This new approach will next permits to add any new variable transforms.

One part of the variable transform is still hardwired: the update of the burp and sqlite files. This will be tackled in the following commit.