Issue #247: Generalization of the observation transform - part 2
The update of the flags (copying the assimilation flag of the transformed variables to the source variables) has been generalized.

As for the updata of the output file, it was found that no modification is needed for the sqlite files (Wow!).

However, signgificant changes are still needed for the burp files and will be tackled next.