Issue #249: Add SubSample for initializing EPS
Now a separate set of ensemble analyses and increments can
be produced for initializing medium-range ensemble forecasts.
This includes adding a potentially different amount of
additive inflation and recentering the smaller ensemble on
the mean of the full ensemble.

Also, minor changes were made to allow filtering the radiance
channels according to the "util" column of the stats_tovs file.