Issue #247: Bugfixes
Three programs were not compiling followinf the latest modifications (addIncrement, obsIO and prepCMA) due to missing librairies in $LIBAPPL in src_file/compile_setup.

Also, the unitTest of obsIO and prepCMA were not validating because obsFilter was not initialized.

Note that is was decided to not modify the update of the burp files because this code is too complex and obscure. Therefore, the users using burp files and wanting derived obs values (including o-p and o-a) in the output files will need to activate the diagnotic sqlite outputs as the output burp files will always contains the same obs than the input files EXCEPTED for u-wind and v-wind.