Issue #260: Squashed commits for prepcma changes
Issue #260: Include suprep

The suprep subroutine set the assimilation flag based on
the selected bufr elements and the list of rejection flags
as specified in the namelist. The updated results for the
prepcma test can be found in the directory:

Issue #260: Reject observations outside time window

The time module namelist "namtim" specifies the length
of the assimilation window (6hours). Then the obsTimeInterp_mod
module routines are used to reject observations outside the
time window by setting the assimilation flag to zero.

No observations in the system test are outside the 6h time
window, therefore the results of the test are not changed.
However, the code was tested with a shorter time window
to ensure that the observations outside this shorter window
were correctly rejected.

Issue #260: Rejection of IR obs near poles

Issue #260: Set obs error and modify for AMSUB

Obs error stddev is set for all observations, then the values
for AMSUB in the tropics are modified.

The channels indicated for assimilation in the stats_tovs files
(i.e. in the "util" column) are used to set the assimilation flag
(i.e. obs_ass).

Issue #260: spatial thinning and cleaning of obsSpaceData

With new logical namelist variables "obsClean" and "thinning",
horizontal thinning and "cleaning" of the observations can be
performed. The thinning procedure was modified slightly so that
it can be performed before the "cleaning" while still producing
the same results. This is necessary to allow the writing of burp
files (still to come).

Issue #260: write to burp files, also fixed an earlier bug

The bug was due to reapplying suprep after some other checks
were done, such that the assimilation flag was incorrectly reset.

Issue #260: small bug fix and simplify head/bodySet_r

All of the system tests were run and some did not pass
due to not accounting for "obs_real" correctly. Some tests
ran MUCH slower than before, due to using "real(x,obs_real)"
in calls to obs_bodySet_r... which is strange. Therefore, the
approach was changed by making headSet_r and bodySet_r
compatible for both single and double precision input
arguments. This simplifies the code and also regained
the previous performance level.

Issue #260: Clean up of regions_mod to remove unneeded code

The `regions_mod` module is now closer to following MIDAS
coding standards and all unused/unneeded routines were
removed to avoid confusion (one of the routines removed does
a "sort" of obsSpaceData, which would mess up the updating of
burp files in MIDAS, so safer to remove).

Also, the `openfile` routine was removed from prepcma and
replaced by `fnom`.

Issue #260: Move rejectHighLatIR from prepcma into enkf_mod

This move allows the routine to be called from both the
prepcma and letkf programs.

Issue #260: Fixes after rebase