Issue #249: Enable Yin-Yang analysis, add global system test
The weight interpolation needed to be improved to be able
to perform analysis directly on Yin-Yang grid. A large global
Yin-Yang system test was added (uses 36 nodes on xc40).

To save some memory the ensemble HX values in the
ensembleObservations module were changed from real8 to
real4.

Improvement to writing of reference P0 to ensemble increment
files so that it is now done in parallel (was taking too much
time for large ensemble).