maestro 'check' task: Show a summary of different between BURP files
When comparing BURP files, it happens that many scattered tiny
differences appear which makes very difficult and tedious to inspect.
Here, we use 'git diff' to extract the minimal difference on each line
and a little Python script to compute the difference.  This is giving a
summary of the differences which greatly facilitates the inspection of
differences.