Issue #258: Many changes to system tests and compilation
The reference results were changed to point to the latest
master branch of the results. Also, changes were made to
the resources to make all tests run. Several changes also
made to remove any reference to specific machine names
throughout the maestro suite and the compilation programs.
Currently, the machine names are only specified in resources.def,
including machines used for storing reference results and for
writing updated reference results. The goal is to make
everything compatible on both new and old machines. But it
has not yet been tested on the old machines.