Issue #267: Allow LPR analysis variable, convert to/from PR
LPR can now be an analysis variable in which case PR will be
read from trial file and converted to LPR. When writing the
analysis and increment (rehm) files, LPR is converted back to
PR. However, LPR is written to the original increment file
(rebm). This is done in the increment_mod module. The same
conversions are also done when writing the analyses and
increments in the letkf program.

The same procedure is applied to LVIS/VIS, which is slightly
different from how it is currently treated.

To simplify, namelist variables were eliminated in the letkf
program to control the writing of increments and the updating
of all members such that increments are always written and all
members are always updated.