Issue #266: Bens now functional for analysis on TH or MM only levels
A changes was also made to allow the user to specify the date of validity of the ensemble forecast trgough the new namelist variable ensDateOfValidity.

When ensDateOfValidity = -1, the data of validity is not taken into account which is usefull for the cases where an ensemble of lagged forecast differences (poor-man EnVar) is used.