Issue #271: Addition of the lake interpolator
- The main change is filling the subroutine s2c_setupLakeInterp with
code, which was only a place holder up until now.

Other minor changes include:

-the switch from the variable name "codtyp"
to "codeType" to be consistent with its use elsewhere in the code.
One benefit is to be able to find that variable in the code using "grep".

- A comment is added in the function s2c_getFootprintRadius for the use
of the nearest neighbor interpolator, in that case the returned footprint
radius is -2. Also in that same function, the lake ice data is now
identified using the codeType 178 instead of relying on the station id,
as is now also done in the module obserrors_mod.f90.

- In case the observation is rejected in subroutines s2c_setupBilinearInterp,
s2c_setupNearestNeighbor, and s2c_setupFootprintInterp, the constant
parameter "obs_notAssimilated" is used instead of the hard coded "0".