Issue # 274: Add a 2D field of correlation length scale and background std for Bdiff
This issue contains two modification:

1) a 2D field of correlation length scale for the B matrix computed using
the diffusion operator is read from the bgstdev file;

2) the same file bgstdev contains also a 2D field
of the background standard deviation.
The difference with previous versions of the code is that
the 2D field may be on any grid (U, Z or G) now.
The interpolation of the field onto the analysis grid
is done using gsv_readFromFile routine.