Issue #288: The script 'collect_unittest_listings' can save listings in any directory
Previously, the script 'collect_unittest_listings' was saving the
listings just besides the reference files.  Since the runner is run as
another user, then the gitlab runner cannot copy the listings in that
directory.  By speciffying the environment variable
'collect_unittest_listings_output', the script will save in this
directory the listings in a directory structure similar to the
unittests.