Issue #290: Importing bgck.satqc_atms program in MIDAS.
'satQCATMS' is the duplicate of bgck.satqc_atms and the does not
use any module for now. Putting different s/r of this program in
modules would cause major revision of the program. This task was
postponed to future issue to make the least changes and facilitate
comparing the results against the results from bgck.satqc_atms.