Issue #291:Observation errors for GPSRO
Basically it is a change of the line 1067 of the oer_fillObsError subroutine to adapt the size of that error the obs of GPSRO of the case of bending angle asimilated. There are no changes in the test results (/ home / dlo001 / midas-291) and the change is necessary for the assimilation of bending angle in cycles, without the change we have verified that the cycle can be stopped by finding a value with a size out of range.