Issue #291: Observation errors for GPSRO
Now only conv_ro_bending does not pass the check (home/dlo001/midas-291), because we have changed the nml of the test to have the EOR column in the output burp. Before, we avoided the  error of the range for two reasons, bending angles were not assimilated therefore we did not detect the error of the range in the oer_fillObsError subroutine and also we did not write the EOR column in the output burp avoiding that Midas stops for having an error out of range.

Therefore it is the only test that fails due to  the new column and also with cases where the bending angles is out of range (0.08) the resut change but it is not problematic because they are not assimilated observations.