Issue #290: Adding satQCAtms s/r to bgckmicrowave_mod.
The s/r used in satQCAtms program are transferred from
bgcksatqc_mod to bgckmicrowave_mod. bgcksatqc_mod is not used
anymore. For this transition, some of the new module's public
variables were already declared in the bgckMW and inovQCAtms and
used as local variables. These variables are brought inside the
module and removed the declaration in the programs.

Re-directing the ASCII output of the inovQCAtms to the listing
rather than to out.txt.