Issue #290: Remove some parameters/arrays from module public variables.
The arrays that are removed from module public variables are passed
as input arguments between different s/r. The interface of some s/r
are changed to get/pass these variables from/to the main program.

The parameters that are removed from module public variables are
defined in both the program and the module.