Issue #292: Change to 'ssh sanl000@...' command to be more explicit
Before, it was supposed that the underlying command to be executed by
the user is 'ssm_publish' so the 'ssh' command was including only
arguments needed for 'ssm_publish' which be may confusing.  We changed
the authorized key to run a more generic command,
'/home/sanl000/bin/ssh_command', so when we call the ssh command, we see
explicitely the 'ssm_publish' command that will be done.