Issue #276: Improve diagFiles and remove perturbation mean
The sqlite diagFiles are now quite different in terms of file
names and columns written. They should now be close to ready for
use in LETKF experiments, allowing detailed diagnostic comparisons
with the EnKF.

Also, the ensemble mean analysis perturbation is now removed when
using one of the LETKF algorithms with cross validation.