Remove the host from the path to reference files.
These days, there is no need to use a path like
'${REFERENCE_MACHINE}:${PATH_TO_REFERENCE_FILE}' since the disk space on
which those files are stored are visible from anywhere.  Moreover, when
'${REFERENCE_MACHINE}' is down, then we just can't run the test because
the path explictly mentions the hostname but the files are still
available anywhere using the direct path.

So we remove the host from the path using only
'${PATH_TO_REFERENCE_FILE}'.