Issue #290: inovQCAtms uses the output of satQCAtms directly.
The unnecessary writing to the BURP between satQCAtms and
inovQCAtms is removed and inovQCAtms uses the output of satQCAtms
directly (instead of reading from the BURP output of satQCAtms in
the previous version). This means there is only one BURP input to
the program and it generates only one output BURP.

The common variables between the two programs keep the name used by
'satQCAtms' (first) program. The variables needed by the inovQCAtms
that did not exist in the satQCAtms program keep the name used by
inovQCAtms program and are added to mwbg_getData s/r to be read
from the input BURP file.

The tasks done before processing of TOVS in inovQCAtms, reading the
ASCII stats_atms_assim, and reading MF/MX fields, are moved before
start reading the BURP in the original program.

Changing the header from DERIALT->BGCKALT, which was done in the
original inovQCAtms program, is moved to the proper place after
scanning of the BURP file is finished.

New s/r in bgckmicrowave_mod, mwbg_updatFlgAtmsF90, which uses
the F90 BURP interface is created to update the flags in inovQCAtms
posrtion of the program. The original s/r was mwbg_updatFlgAtms,
where the F77 BURP fuctions were used.

Proper spacing added between s/r input arguments. The list of the
arguments had to be split between multiple lines in some cases.

The description of original inovQCAtms program is added before the
section for declaration of variables in the main program.