Issue #290: Removing the satQCAtms/inovQCAtms unitTests.
The two programs are combined and tested in bgckAtms and separate
unitTests are not needed. Therefore, satQCAtms/inovQCAtms unitTests
are deleted.

In bgckmicrowave_mod, the s/r mwbg_updatFlgAtms (using F77 functions)
is removed since it was only used in original inovQCAtms. The
mwbg_updatFlgAtmsF90 in renamed to mwbg_updatFlgAtms.