Issue #294: Bug Fix -- Add missing O-P check for surface ES observations
The missing section in function isetflag for setting the background O-P check flag for surface
dewpoint depression (ES) observations (BURP element BUFR_NESS) is added.

In addition, another small bug is fixed in the same function (isetflag) where, for most
observation types, the function will return a flag value of 0 (no rejection) if the value
being checked is exactly equal to the level 2 rejection criteria. Also all the comparison
operators like ".ge." are changed to the more modern style e.g. ">=".

Note that this modification will affect the results of the background check (for conventional
observations) thereby affecting the filtering of observations entering the analysis. In essence,
the number of observations selected for assimilation will be reduced.