Issue #297: Re-ordering updating blocks in BURP to get same results.
The deleting/writing blocks in BURP report will change the order of
blocks and the results will not be identical to reference. To avoid
this issue, updating the block in BURP is done in the same order as
the block appear in input BURP.

Because of the same issue, the updating of FLAG and TERRAIN TYPE
are done in one s/r (updating the block will change its order in
the report). Before the TERRAIN TYPE update was done in
mwbg_ADDTRRNF90 s/r, which is not deleted.

If all the elements in block are set using BURP_Set_Tblval, in
BURP_Write_Block, CONVEERT_BLOCK should be set to FALSE. Otherwise
the output results will change.