Issue #256: Setting the state-dependent observation error in obsSpaceData.
The fit parameters to define the state-dependent observation error
are read from ASCII file, with similar formatting as stats_tovs.
Using the state-dependent observation error is triggered with namelist
variable. The fit parameters together with CLW are used to compute the
state-dependent observation error and set OBS_OER body in obsSpaceData.