Issue #297: Writing the averaged CLW to the BURP.
Averaged CLW is calculated by taking the mean of retrieved CLW from
observation and simulated observation. The simulated observation is
constructed by taking the difference of O' and O'-P (omp). The
averaged CLW is written to the BURP at update stage. Averaged CLW
is the quantity to be used to define the OBS_OER for each observation
in all-sky mode.