Issue #295: New version that does not require changes to namelists.
The modifications were re-done so that there is no longer a need to modify
the namelist files. There is still a separate list of BURP elements for
GB-GPS (GP) family files that can be specified in the NAMBURP_FILTER_SFC
section but if the list is not there, a default list is used.