Adding CLW to the nl/tl/ad rttov obsOperator.
CLW is used to run the obsOperator for a list of MW instruments
indicated in the NAMMW namelist block.

A new function, isInstrumMicrowave, is created to determine if the
instrument is MW.

A minimum value of 1.D-9 is used for the extrapolated CLW to be used
in RTTOV profile for nl obsOperator.