Issue #256: Use NAMTOV namelist to control which instruments use CLW in nl/tl/ad.
In NAMTOV, list of instruments to use CLW is given. If no name is
provided, the CLW is not used. The 'isInstrumIdInTheList' function
loops through this list and determines if the given instrument is
in the list or not.

CLW is used in nl/tl/ad for a specific instrument only if:
  1- It is available in the columnData object.
  2- At least one instrument to use CLW is prescribed in the NAMTOV
        namelist.
  3- Using clw_data is activated for that sensor.
  4- Additional namelist variable, mwInstrumentUsingCLW_tl, controls
        if CLW is used in the tl/ad RTTOV obs operators.

The allocation/interpolation of CLW profiles, and populating the
rttov_profile structure are done only if the above criteria are met.