Issue #298: Using only a single variable to identify the maximum execution time allowed.
We had previously an hard-to-understand logic which needed to specify a
mean execution timing and a maximum expressed as a percentage of the
mean.

We change that with a single number expressing the maximum execution
allowed using the variable 'UnitTest_maximum_execution_time'.  When the
execution time exceeds this value then a message is issue or it aborts
if 'UnitTest_abort_when_over_maximum_exection_time=abort'.