Issue #256: Properly initialize the variable to allow running RTTOV using CLW.
In RTTOV non-linear/linear modules, for each sensor the
runObsOperatorWithClw* variable is initialized to .false. to prevent
using CLW in RTTOV if not requested.

Adding the abort in non-linear RTTOV module if instrument is set to
use CLW in the namelist but LWCR is not in NAMSTATE.

Adding QC in the list of 3D analyzed variables.