Issue 246: New OmP stddev element in obsSpaceData
The new element OBS_OMPE has been added to obsSpaceData
to hold the OmP error std dev for use in background check.
The background check step has also been modified to use OBS_OMPE,
when the OmP error std dev is available, instead of the combination
of OMPS_OER and OMPS_HBHT for identifying if the obs will be
assimilated. While the process of allowing for use of OBS_OMPE
has been implemented for all conventional and retrieved chemical
obs, the routines for providing values to OBS_OMPE have only been
implemented for retrieved chemical obs. Related changes would be needed
for the conventional (and possibly BT) obs.

Two possible approaches have been introduced for providing the OmP
error standard deviations of retrieved chemical obs. Both rely on an
input text file (the same as the one containing information on the
obs error std dev) directing the code on the action to take. One approach
is to use the OmP error standard deviations provided by the text
file. The other approach is to calculate the OmP error standard deviations
from the data of the current 6h interval on the condition that the
dataset is large enough for each selected latitude bin. The latter is similar
to what is currently done for the RO and GP obs families.

Some indenting and formatting updates have also been done.