Issue #335: First attempt to compute O-P, still some bugs
The O-P is computed, but not yet validated. Before being able
to validate, first the bias correction must be introduced and
then a latent bug related to the array `rejectionCodArray`
needs to be fixed. This bug is due to incorrect dimensions
and causes an abort when the debug mode is activated.