Issue #332: Adding ASCAT for sea ice analysis
A call to the new subroutine filt_iceConcentration is added to select
only observations from satellites specified in the namelist NAMFILT.

Because backscatter anisotropy measure becomes abiguous for distinguishing
between ice and water where wind speed is too low, a call to the new
subroutine filt_bkscanis is added to filter out ASCAT data.

Also a call to the new subroutine oer_SETERRBKSCANIS is added to set
observation-error standard deviations for backscatter anisotropy, which
depends on the background ice concentration.