Issue #332: Addition of ASCAT data for sea ice analysis
A new SQLite schema for ASCAT data is added (i.e. gl_ascat).
In this case new HEADER columns track_cell_no and mod_wind_spd are read.
The track_cell_no in SQLite files goes from 1 to 42, 21 cells from
each side of the satellite ground track. The track_cell_no is converted
to a number between 1 and 21 assuming a symmetry from left to right of
the satellite ground track. This is the assumption used during the regression
training between backscatter anisotropy and sea ice concentration.