Issue # 296
Tools to represent GPSRO vertical correlation within a profile.
The correlation is expressed through a list of 4 ansatz functions.
For each satellite, called by its satid,  (ultimate cause: the processing method),
correlation is expressed as a list (WGPS) of contributions from these ansatz. The
lists are expressed through namelist arrays.
Besides the (small) changes in code, the default was set to give no
weight to data not explicitly mentioned in the GPSRO namelist.
This required modification of numerous namelists from the unit tests, to give all
participating data a unit weight, which recovers backwards
compatibility.
In a usual setting, weights should not be 1, but rather of the order of 0.2. Details TBD
from correlation analysis. Some initial fits to these ansatz have
already been tested.