Issue # 296
Eliminated array IGPSSAT, its function substituted by the presence of a non-zero WGPS.
This simplifies GPSRO namelist.
The namelist still requires a non-zero WGPS(satid,1:4) for each satid
(satellite id) to be considered, or data will be rejected.
For mere testing purposes, the following is sufficient
WGPS(satid, 1:4) = 1., 0., 0., 0.,
Proper operational coefficients in WGPS to be determined by a correlation study.