Issue #235: Reintroduce ability to compute Z and P on columns
This is reintroduced only for the increment (s2c_tl and s2c_ad)
as a way of reducing mpi communication in the affected
subroutines. As long as the increment is interpolated to
a vertical column (i.e. not slant path), then this is a
good option to save time. A new namelist variables is
used to activate this "new" feature:

`calcHeightPressIncrOnColumn=.TRUE.`.