Issue #357: tovsFilt now mpi, results not sensitive to topology
Arbitrarily, when multiple observations are the same distance
from the grid point, the one with minimum longitude is taken
to remove any ambiguity. This should mean that results are
insensitive to either the mpi topology or the order of the
observations in the input file.

The check on duplicate observations is still not mpi, but
this doesn't matter for the files in the current system
test, since no duplicate obserations are present.