Issue #362: We will run 'splitobs' CI automatic tests only when a tag is pushed.
Since those tests are not yet parallelized, they take more than one hour
to run.  Also, the code for 'splitobs' is separated from the rest f the
code base and it is not likely to change often so it is not worth to run
the tests on each merge into 'master' branch.  But it is certainly worth
running them when a tag is pushed.