Issue #372: Bugfix to use symmetric CLW for QC test 12 in all-sky mode.
CLW from observation is used for QC in clear-sky mode. In all-sky
mode, the symmetric CLW should be used both to inflate the obs
error for QC test 14 and do cloud liquid water QC test 12. So far,
the CLW from observation was used to perform QC test 12 in
all-sky mode.

Using the symmetric CLW for QC test 12 is activated by a namelist
variable, default is set to .false. for backward compatibility
with clear-sky assimilation. If using state-dependant
obs error for all-sky assimilation and the using symmetric CLW is not
activated for QC test 12, the code aborts. The code also aborts if
using symmetric CLW is activated in clear-sky mode.