Issue #372: Use clear radiances in new QC test of bgckMW in all-sky mode.
Clear-sky radiances are used in new QC test 16 of bgckMW for all-sky
assimilation.

In this test, the clear and cloudy radiances of channel 5 are used to
compute simulated and observed cloud effects. If either of these
quantities is smaller than -0.5K, channels 4-5 are rejected over water
to exclude radiances affected by extreme scattering.