Issue #374: bug in midas.reunir_obs_mpi
Bug: loss of RESUME and RDB4_SCHEMA content in output files.
Fix: for both if blocks deciding wether or not to copy the info into new file, the line changing the status from 0 to 1 was before the if on status -eq 0, hence preventing it from entering the if block and copying the table's content. The line was cut and pasted at the end (inside) the if block, so that the info is copied once, as intended. This was done for the if on RESUME and the one on RDB4_SCHEMA.