Issue #367: System test for "conv" and change copying
It was also necessary to change the strategy a little
for creating the before and after thinning obs files.
Now the following steps are followed:
* original files are copied from `obs` to `obsOriginal`
* the files in `obs` are updated after the background check
* these updated files are copied to `obsBeforeThinning`
* the files in `obsOriginal` are copied to `obs`
* the files in `obs` are updated afte the thinning

This was necessary to get results that look much closer
to the operational bgckalt files, especially for radiosondes
(even though no thinning is done yet for them).