Issue #372: Bugfix for CLW used to define obs error in bgck.
The averaged CLW used to define the symmetric obs error should be
between 0 and 3. For some locations (e.g. abs(lat)>50 over sea-ice),
the clw_avg is set to missing -99.0 value in Grody s/r. This missing
value can not be used to define the obs error (otherwise, the
clear-sky obs error is assigned to this observation, as
clw_avg<clwThresh1). All channels that need CLW for their observation
error at this location are now rejected in all-sky mode in test 12.

In all-sky mode, O-P check (test 14) is skipped when the clw_avg is
missing. The observation is already rejected in test 12.