Issue #372: obsSpaceData CLW to inflate obs error at analysis.
Element 13209 of BURP is clwObs and is copied to OBS_CLW1 header
column of obsSpaceData. Element 13109 of BURP is clwFG and is copied
to OBS_CLW2 header column of obsSpaceData.

Changes to brpr_readBurp to read in the element 13109 from BURP
file.

OBS_CLW1 and OBS_CLW2 are used to inflate the observation error
at analysis step.