Issue #385: New home module for the getPositionXY function
The getPositionXY function was previously in the utilities_mod module,
but had to be moved in this new interpolation_mod module to resolve
a dependency issue caused my the use of kdtree. Hence, the function was
renamed from utl_getPositionXY to int_getPositionXY.

A second important change is the use of ezgprm and gdll instead of
gdllfxy. The lat-lon positions are kept in memory and are used in
subsequent calls to the subroutine, resulting in a significant
speed-up of the program.