Issue #385: Major redesign of the module for ORCA grids.
A new approach is used for the ORCA grids family. Instead of searching
for the nearrest 4 grid points forming a "square" around the reference
point amongst the locations returned by kdtree2, only use the nearest
grid point returned by kdtree2. Then a perturbation method around this
nearest grid point is used to find the 4 grid points needed to
calculate the real x and y position on the grid.