Issue #395: Pass RO bending blocks to output
In order to preserve, for downstream use, the shape of the observation
column, necessary for slant code, this shape must be present in a data
block that is passed to output. An initial idea was to preprocess the
input BURP to locate this shape in the assimilated block.

Instead, it was here implemented a modification to the burp update,
which forwards the module that contains this shape. It was previously
passed to output only if it contained assimilated observations. It is
now forwarded unconditionally. We also forward its marker block, as well
as a small miscellaneous block that contains orbit information. It may be
useful in the future, at least for statistical purposes, and perhaps to
allow more advanced operators.

Only burpread_mod.f90 is modified, and the result is backwards bit
compatible in the functional sense. The check fails, though, as new blocks
are present in the output burp files. Thus new reference output are issued
for the BgckConv, conv and conv_ro_bending Tests.