Issue #382: Using '/home/erv000/data/ords/SSM/rttov-12/hugepages16m-1.5-1-g6f34544'
The difference with 'eccc/mrd/rpn/anl/rttov/12v1.5' is that, when
compiling the library, we should have load the compiler before the
code-tools.

When loading the code-tools before the compiler, it happens to set the
compiler rules for the default compiler which is the GNU compiler.  This
was affecting the results.