Issue #382: Using 'eccc/mrd/rpn/anl/rttov/12v1.5.1'
The difference with 'eccc/mrd/rpn/anl/rttov/12v1.5' is that, when
compiling the library, we should have load the compiler before the
code-tools.

When loading the code-tools before the compiler, it happens to set the
compiler rules for the default compiler which is the GNU compiler.  This
was affecting the results.