Issue #399: Programs abort if `huMinValue` not in namelist
This applies to all program that use the ensemble B matrix.

The abort will only happen if HU is an analysis variable and the
control variable for the ensemble B matrix is LQ. This ensures
that the use must specify the value of `huMinValue` that they
want to use, avoiding the accidental use of the wrong value.