Issue #428: For more security, touch files before launching MPI
We did touch the file in the MPI job itself, but it the MPI launch
itself does not work then the script would not detect the problem.  By
creating the 'touch' files before any MPI is involved and the each MPI
task is erasing its own 'touch' file when its job has been completed
then we are sure that everything went well.